+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # demo
 
+## 1.0.69
+
+### Patch Changes
+
+- Updated dependencies [56957d2]
+  - @xaui/icons@0.0.12
+  - @xaui/native-legacy@0.2.10
+
+## 1.0.68
+
+### Patch Changes
+
+- Updated dependencies [f455c83]
+  - @xaui/native-legacy@0.2.9
+
 ## 1.0.67
 
 ### Patch Changes